refactor(frontend): consolidate layout, landing, and display components
