+++ frontend/src/components/BatchResultsDisplay.test.tsx
