+++ frontend/src/components/BatchResultsDisplay.tsx
