+++ frontend/src/components/ShowcaseSections.tsx
